+++ spring-data-neo4j-rx/src/main/java/org/neo4j/springframework/data/repository/query/PartTreeNeo4jQuery.java
@@ -18,15 +18,12 @@
  */
 package org.neo4j.springframework.data.repository.query;
 
-import static java.util.stream.Collectors.*;
-
 import java.time.Instant;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.OffsetTime;
 import java.time.ZonedDateTime;
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.EnumSet;
 import java.util.HashSet;
@@ -36,7 +33,6 @@ import java.util.Set;
 import java.util.function.BiFunction;
 
 import org.neo4j.driver.Record;
-import org.neo4j.driver.types.Point;
 import org.neo4j.driver.types.TypeSystem;
 import org.neo4j.springframework.data.core.Neo4jOperations;
 import org.neo4j.springframework.data.core.PreparedQuery;
@@ -45,9 +41,7 @@ import org.springframework.data.repository.query.RepositoryQuery;
 import org.springframework.data.repository.query.parser.Part;
 import org.springframework.data.repository.query.parser.PartTree;
 import org.springframework.data.repository.query.parser.PartTree.OrPart;
-import org.springframework.data.util.ClassTypeInformation;
 import org.springframework.lang.Nullable;
-import org.springframework.util.Assert;
 
 /**
  * Implementation of {@link RepositoryQuery} for derived finder methods.
@@ -113,82 +107,5 @@ final class PartTreeNeo4jQuery extends AbstractNeo4jQuery {
 			.withParameters(queryAndParameters.getParameters())
 			.usingMappingFunction(mappingFunction)
 			.build();
-	}
-
-	/**
-	 * Checks whether the given part can be queried without case sensitivity.
-	 *
-	 * @param part query part to check if ignoring case sensitivity is possible
-	 * @return True when {@code part} can be queried case insensitive.
-	 */
-	static boolean canIgnoreCase(Part part) {
-		return part.getProperty().getLeafType() == String.class && TYPES_SUPPORTING_CASE_INSENSITIVITY
-			.contains(part.getType());
-	}
-
-	static class PartValidator {
-
-		private final Neo4jQueryMethod queryMethod;
-
-		PartValidator(Neo4jQueryMethod queryMethod) {
-			this.queryMethod = queryMethod;
-		}
-
-		void validatePart(Part part) {
-
-			validateIgnoreCase(part);
-			switch (part.getType()) {
-				case AFTER:
-				case BEFORE:
-					validateTemporalProperty(part);
-					break;
-				case IS_EMPTY:
-				case IS_NOT_EMPTY:
-					validateCollectionProperty(part);
-					break;
-				case NEAR:
-				case WITHIN:
-					validatePointProperty(part);
-					break;
-			}
-		}
-
-		private void validateIgnoreCase(Part part) {
-
-			Assert.isTrue(part.shouldIgnoreCase() != Part.IgnoreCaseType.ALWAYS || canIgnoreCase(part),
-				() -> String.format(
-					"Can not derive query for '%s': Only the case of String based properties can be ignored within the following keywords: %s",
-					queryMethod,
-					formatTypes(TYPES_SUPPORTING_CASE_INSENSITIVITY)));
-		}
-
-		private void validateTemporalProperty(Part part) {
-
-			Assert.isTrue(COMPARABLE_TEMPORAL_TYPES.contains(part.getProperty().getLeafType()), () -> String
-				.format(
-					"Can not derive query for '%s': The keywords %s work only with properties with one of the following types: %s",
-					queryMethod, formatTypes(Collections.singletonList(part.getType())),
-					COMPARABLE_TEMPORAL_TYPES));
-		}
-
-		private void validateCollectionProperty(Part part) {
-
-			Assert.isTrue(part.getProperty().getLeafProperty().isCollection(), () -> String
-				.format("Can not derive query for '%s': The keywords %s work only with collection properties",
-					queryMethod,
-					formatTypes(Collections.singletonList(part.getType()))));
-		}
-
-		private void validatePointProperty(Part part) {
-
-			Assert.isTrue(ClassTypeInformation.from(Point.class)
-				.isAssignableFrom(part.getProperty().getLeafProperty().getTypeInformation()), () -> String
-				.format("Can not derive query for '%s': %s works only with spatial properties", queryMethod,
-					part.getType()));
-		}
-
-		private static String formatTypes(Collection<Part.Type> types) {
-			return types.stream().flatMap(t -> t.getKeywords().stream()).collect(joining(", ", "[", "]"));
-		}
 	}
 }
